### 2023-06-09 #5528 @editorialbot
[PRE REVIEW]: Bramble: adaptive common neighbor analysis (CNA) for the recognition of surface topologies in nanoparticles
**Submitting author:** @ifilot (<a href="http://orcid.org/0000-0003-1403-8379">Ivo Filot</a>)
**Repository:** https://github.com/ifilot/bramble
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
