+++ docs/index.rst
@@ -9,6 +9,8 @@ Bramble: Fast Common Neighbor Analysis
    :target: https://www.gnu.org/licenses/gpl-3.0
 .. image:: https://joss.theoj.org/papers/f6493d619d92bb6d993713b9d1abb38c/status.svg
    :target: https://joss.theoj.org/papers/f6493d619d92bb6d993713b9d1abb38c
+.. image:: https://zenodo.org/badge/641289155.svg
+   :target: https://zenodo.org/badge/latestdoi/641289155
 
 :program:`Bramble` is a single atom pattern recognition algorithm based on the
 :ref:`Common Neighbor Analysis method <background>`. It can efficiently construct
