### 2023-07-31 #5710 @editorialbot
[REVIEW]: Bramble: adaptive common neighbor analysis (CNA) for the recognition of surface topologies in nanoparticles
**Submitting author:** @ifilot (<a href="http://orcid.org/0000-0003-1403-8379">Ivo Filot</a>)
**Repository:** https://github.com/ifilot/bramble
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.2
