### 2023-08-11 #5710 @hsulab
Hi @ifilot,

Bramble is an analysis tool that can detect different active sites on metal nanoparticles. I think it will be quite useful and find broad applications in computational heterogeneous catalysis.

# Installation

It took me a lot of time to install it. Since I tested it on a remote machine (Intel Skylake and Nvidia A100), I used conda to install dependencies

- conda install -c pkgs/main gcc_linux-64=9.3.0 gxx_linux-64=9.3.0
- conda install -c conda-forge tclap
- conda install -c conda-forge boost

and manually installed the eigen3 v3.4.0.

The version of cudatoolkit is 12.2.

For the CPU version, I have to make changes to CMakeLists.txt as

```
-find_package(Boost COMPONENTS regex unit_test_framework REQUIRED)
+find_package(Boost COMPONENTS regex unit_test_framework system REQUIRED)

-                    ${TCLAP_INCLUDE_DIR}
+                    ${TCLAP_INCLUDE_DIRS}

```

The compilation works with eigen 3.3.4, 3.3.7, and 3.4.0 but it failed with g++ 8.2 related to `std::filesystem` and g++ 11.2 `error: call of overloaded 'sqrt(double)' is ambiguous`.

For the GPU version, I also changed the CMakeLists.txt as

```
SET(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -DMOD_CUDA -no-pie")
```

Otherwise, it will have errors like `relocation R_X86_64_32 against '.rodata' can not be used when making a PIE object; recompile with -fPIE`. Meanwhile, the GPU version only works with eigen 3.4.0. I tested it with eigen 3.3.4 and 3.3.7 but there is `fatal error: math_functions.hpp: No such file or directory`.

Also, I wonder if **bramblecuda** is only used for showing the GPU information and the actual computation is still performed by **bramble**.

# Application

- I tested the first example and did not finish the second one as it took too long. The outputs by cpu and gpu versions are shown below. I did not observe any acceleration by GPU. I wonder if my GPU version was not compiled successfully. I think the author should add some benchmarks in the documentation.

```
--------------------------------------------------------------
Executing Bramble v.1.0.0
Author: Ivo Filot <i.a.w.filot@tue.nl>
Documentation: https://bramble.imc-tue.nl
--------------------------------------------------------------
Compilation time: Aug 10 2023 00:13:38
Git Hash: 8c05d47
--------------------------------------------------------------

Loading src/test/data/POSCAR_Rh111 as POSCAR/CONTCAR file.
Geometry file contains 7 atoms.
Calculating interatomic distances...
Completed calculation of interatomic distances.
Start Similarity Analysis
Copying interatomic distances to similarity analysis object.
Constructing local density matrices... Done!
Constructing jobs and pre-allocating permutation vectors.
Building permutations: 9
Building permutations: 12
Executing similarity analysis using 80 threads.
Jobs remaining: 0
 100% [|||||||||||||||||||||||||||||||||||||||||||||||||||||]

--------------------------------------------------------------
Done in 281.107 seconds.
```

```
--------------------------------------------------------------
Executing Bramble v.1.0.0
Author: Ivo Filot <i.a.w.filot@tue.nl>
Documentation: https://bramble.imc-tue.nl
--------------------------------------------------------------
Compilation time: Aug 11 2023 00:26:51
Git Hash: bc1a8d7
Modules: CUDA
--------------------------------------------------------------

Loading ./test/POSCAR_Rh111 as POSCAR/CONTCAR file.
Geometry file contains 7 atoms.
Calculating interatomic distances...
Completed calculation of interatomic distances.
Start Similarity Analysis
Copying interatomic distances to similarity analysis object.
Constructing local density matrices... Done!
Constructing jobs and pre-allocating permutation vectors.
Building permutations: 9
Building permutations: 12
Executing similarity analysis using 80 threads.
Jobs remaining: 0
 100% [|||||||||||||||||||||||||||||||||||||||||||||||||||||]

--------------------------------------------------------------
Done in 278.382 seconds.
```

- There is an error when reading pattern2.json.
```
Running bramble-tool v.1.0.0
Author: Ivo Filot <i.a.w.filot@tue.nl>
Documentation: https://bramble.imc-tue.nl
--------------------------------------------------------------
Compilation time: Aug 10 2023 00:13:43
Git Hash: 8c05d47
--------------------------------------------------------------
A problem occurred reading the pattern file.
Error: Pattern "FCC Bulk" is invalid!
```

# Miscellaneous

* Community guidelines are not provided.

All the best,
Jiayan
