+++ .github/workflows/build.yml
@@ -7,9 +7,9 @@ on:
     branches: [ "master", "develop" ]
 
 jobs:
-  build:
+  build-ubuntu-22.04:
 
-    runs-on: ubuntu-latest
+    runs-on: ubuntu-22.04
 
     steps:
     - uses: actions/checkout@v3
@@ -21,3 +21,33 @@ jobs:
       run: cd build && make -j
     - name: Perform unit tests
       run: cd build && make test
+
+  build-ubuntu-20.04:
+
+    runs-on: ubuntu-20.04
+
+      steps:
+      - uses: actions/checkout@v3
+      - name: Install dependencies
+        run: sudo apt install -y libeigen3-dev libboost-all-dev libtclap-dev
+      - name: Configure CMake
+        run: mkdir build && cd build && cmake ../src
+      - name: Build
+        run: cd build && make -j
+      - name: Perform unit tests
+        run: cd build && make test
+
+  build-ubuntu-23.10:
+
+    image: ubuntu-23.10
+
+      steps:
+      - uses: actions/checkout@v3
+      - name: Install dependencies
+        run: sudo apt install -y libeigen3-dev libboost-all-dev libtclap-dev
+      - name: Configure CMake
+        run: mkdir build && cd build && cmake ../src
+      - name: Build
+        run: cd build && make -j
+      - name: Perform unit tests
+        run: cd build && make test
