+++ .github/workflows/build.yml
@@ -39,7 +39,8 @@ jobs:
 
   build-ubuntu-23.10:
 
-    image: ubuntu-23.10
+    runs-on: ubuntu-latest
+    container: ubuntu/ubuntu:23.10
 
       steps:
       - uses: actions/checkout@v3
