### 2023-08-24 cc4882a
Adding cmake to install
