+++ .github/workflows/build.yml
@@ -14,7 +14,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install dependencies
-      run: sudo apt-get update && sudo apt-get install -y libeigen3-dev libboost-all-dev libtclap-dev cmake
+      run: sudo apt-get update && sudo apt-get install -y build-essential libeigen3-dev libboost-all-dev libtclap-dev cmake
     - name: Configure CMake
       run: mkdir build && cd build && cmake ../src
     - name: Build
@@ -29,7 +29,23 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install dependencies
-      run: sudo apt-get update && sudo apt-get install -y libeigen3-dev libboost-all-dev libtclap-dev cmake
+      run: sudo apt-get update && sudo apt-get install -y build-essential libeigen3-dev libboost-all-dev libtclap-dev cmake
+    - name: Configure CMake
+      run: mkdir build && cd build && cmake ../src
+    - name: Build
+      run: cd build && make -j
+    - name: Perform unit tests
+      run: cd build && make test
+
+  build-ubuntu-23-04:
+
+    runs-on: ubuntu-latest
+    container: ubuntu:23.04
+
+    steps:
+    - uses: actions/checkout@v3
+    - name: Install dependencies
+      run: apt-get update && apt-get install -y build-essential libeigen3-dev libboost-all-dev libtclap-dev cmake
     - name: Configure CMake
       run: mkdir build && cd build && cmake ../src
     - name: Build
@@ -45,7 +61,7 @@ jobs:
     steps:
     - uses: actions/checkout@v3
     - name: Install dependencies
-      run: apt-get update && apt-get install -y libeigen3-dev libboost-all-dev libtclap-dev cmake
+      run: apt-get update && apt-get install -y build-essential libeigen3-dev libboost-all-dev libtclap-dev cmake
     - name: Configure CMake
       run: mkdir build && cd build && cmake ../src
     - name: Build
