+++ src/CMakeLists.txt
@@ -19,7 +19,7 @@
  #*************************************************************************/
 
 # set minimum cmake requirements
-cmake_minimum_required(VERSION 2.8.12)
+cmake_minimum_required(VERSION 3.5)
 project (bramble)
 
 # add custom directory to look for .cmake files
