### 2023-08-24 ac2c2b9
Fixing incorrect entries in CMakeLists
