+++ src/CMakeLists.txt
@@ -67,14 +67,14 @@ set (BOOST_ALL_DYN_LINK ON)
 
 # Include libraries
 find_package(PkgConfig REQUIRED)
-find_package(Boost COMPONENTS regex unit_test_framework REQUIRED)
+find_package(Boost COMPONENTS regex unit_test_framework system REQUIRED)
 pkg_check_modules(TCLAP tclap REQUIRED)
 pkg_check_modules(EIGEN eigen3 REQUIRED)
 
 # Set include folders
 include_directories(${CMAKE_CURRENT_SOURCE_DIR}
                     ${CMAKE_BINARY_DIR}
-                    ${TCLAP_INCLUDE_DIR}
+                    ${TCLAP_INCLUDE_DIRS}
                     ${Boost_INCLUDE_DIRS}
                     ${EIGEN_INCLUDE_DIRS})
 
