### 2023-08-24 #5710 @NicoRenaud
Hi,
first my apologies for taking so long to  review the paper. Thanks for the reminder @jgostick !

# Functionality

- The installation of the dependencies went very smoothly via `apt-get`
- The functionality mentionned in the documentation and in the paper works well. The code seems to be rather memory demanding as the second example got killed by the OOM killer. Is it expected ?
- I haven't been able to test the GPU performance, I have not GPU available locally.

# Documentation
- In the example, it is not clear how one display the distance matrix. I guess it's with the `scripts/similary_plot.py`. A better documentation of this python script might help new users.'
- It would also be nice to add the blender script used to generate the 3d rendering of the nanocrystals.

# Software Paper
- the paper itself is very clear, nothing to change there.
