### 2023-08-26 ca20395
Adding execution times
