### 2023-08-27 0a9315e
Adding calculation times for Co1121 example
