### 2023-08-27 a00b16f
Add notation execution times
