### 2023-09-04 237bf1d
Implementing multi-GPU procedure
