+++ src/test/test_similarity.cpp
@@ -57,7 +57,7 @@ BOOST_AUTO_TEST_CASE(test_similarity) {
     BOOST_TEST(ans2 == ans3, boost::test_tools::tolerance(1e-7));
 
     #ifdef MOD_CUDA
-    float ans4 = sa.calculate_distance_metric_cuda(dm3, dm4, &permvec[0]);
+    float ans4 = sa.calculate_distance_metric_cuda(0, dm3, dm4, &permvec[0]);
     BOOST_TEST(ans2 == ans4, boost::test_tools::tolerance(1e-7));
     BOOST_TEST(ans3 == ans4, boost::test_tools::tolerance(1e-7));
     #endif // MOD_CUDA
