+++ src/similarity_analysis.cpp
@@ -99,7 +99,23 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
         }
     }
 
-    // give number of threads
+    #ifdef MOD_CUDA	// CUDA SUPPORT
+    int cuda_devices = 0;
+    #pragma omp parallel
+    {
+        int nrthreads = omp_get_num_threads();
+        int threadnum = omp_get_thread_num();
+
+        if(threadnum ==  0) {
+            CardManager cm;
+            cuda_devices = cm.get_num_gpus();
+            std::cout << "Planning job using " << cuda_devices << " GPUs and " << (nrthreads - cuda_devices) << " CPUs" << std::endl;
+        }
+    }
+
+    #else	        // NO CUDA SUPPORT
+
+    int cuda_devices = 0;
     #pragma omp parallel
     {
         int nrthreads = omp_get_num_threads();
@@ -110,6 +126,8 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
         }
     }
 
+    #endif // MOD_CUDA
+
     // show progress bar
     ProgressBar progress(jobs.size(), "");
     std::vector<bool> jobdone(jobs.size(), false);
@@ -117,8 +135,10 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
 
     #ifdef MOD_CUDA
     omp_set_nested(1);
-    #pragma omp parallel for num_threads(2) schedule(dynamic)
-    #endif // MOD_CUDA
+    omp_set_dynamic(0);
+    omp_set_num_threads(cuda_devices+1);
+    #pragma omp parallel for schedule(dynamic)
+    #endif // MOD CUDA
     for(unsigned int k=0; k<jobs.size(); k++) {
 
         // keep track of time per job
@@ -128,6 +148,7 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
         const unsigned int j = jobs[k].second;
         std::vector<unsigned int> permvec(12,0);
         float mhsn = -1;
+        int threadnum = omp_get_thread_num();
 
         // refuse to execute job when number of atoms is larger than 12
         if(this->distance_matrices[i].rows() > 12 || this->distance_matrices[j].rows() > 12) {
@@ -136,11 +157,14 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
         }
 
         #ifdef MOD_CUDA
-        if(omp_get_thread_num() == 0) { // run these on GPU
-            mhsn = this->calculate_distance_metric_cuda(this->distance_matrices[i],
+        if(threadnum < cuda_devices) { // run these on GPU
+            std::cout << "Spawning CUDA: " << threadnum << std::endl;
+            mhsn = this->calculate_distance_metric_cuda(threadnum,
+                                                        this->distance_matrices[i],
                                                         this->distance_matrices[j],
                                                         &permvec[0]);
         } else { // all other jobs are run on CPU
+            std::cout << "Spawning regular: " << threadnum << std::endl;
             mhsn = this->calculate_distance_metric_openmp(this->distance_matrices[i],
                                                           this->distance_matrices[j],
                                                           &permvec[0]);
@@ -281,13 +305,16 @@ float SimilarityAnalysis::calculate_distance_metric_openmp(const MatrixXXf& dm1,
 /**
  * @brief      Calculates the distance metric between two distance matrices on the GPU
  *
- * @param[in]  dm1      distance matrix 1
[…]
