### 2023-09-04 7a053e5
Adding command line arguments for NGPU
