+++ src/CMakeLists.txt
@@ -35,8 +35,8 @@ execute_process(
 
 # prepare configuration file
 SET(VERSION_MAJOR "1")
-SET(VERSION_MINOR "0")
-SET(VERSION_MICRO "1")
+SET(VERSION_MINOR "1")
+SET(VERSION_PATCH "0")
 configure_file(config.h.in ../src/config.h @ONLY)
 
 # Enable release build
