### 2023-09-04 bd6c4bd
Fixing bug in GPU number
