+++ docs/publications.rst
@@ -6,6 +6,10 @@ Publications
 
 The following publications make use of :program:`Bramble`
 
+* *Unraveling the Role of Metal–Support Interactions on the Structure Sensitivity
+  of Fischer–Tropsch Synthesis*, van Etten, M.P.C., de Laat, M.E., Hensen, E.J.M.,
+  Filot, I.A.W., J. Phys. Chem. C, **2023**, 127, 31, 15148-15156,
+  DOI: `10.1021/acs.jpcc.3c02240 <https://doi.org/10.1021/acs.jpcc.3c02240>`_
 * *Enumerating Active Sites on Metal Nanoparticles: Understanding the Size
   Dependence of Cobalt Particles for CO Dissociation*, van Etten M.P.C.,
   Zijlstra B., Hensen E.J.M., Filot, I.A.W., ACS Catal., **2021**, 11, 14,
