+++ docs/examples.rst
@@ -77,6 +77,10 @@ acceleration, we refer to the Table as seen below.
      - 172.58s
    * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads) + RTX 4090
      - 90.18s
+   * - Intel(R) Core(TM) i5-8400 CPU @ 2.80GHz (6 threads)
+     - 311.84s
+   * - Intel(R) Core(TM) i5-8400 CPU @ 2.80GHz (6 threads) + RTX 2070
+     - 125.08s
 
 Co HCP 11-21
 ------------
@@ -193,6 +197,6 @@ acceleration, we refer to the Table as seen below.
    * - System
      - Execution time (averaged)
    * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads)
-     - 2368.63s (39m28s)
-   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads) + RTX 4090
      - 5207.93s (1h26m47s)
+   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads) + RTX 4090
+     - 2368.63s (39m28s)
