+++ .github/workflows/build.yml
@@ -22,12 +22,16 @@ jobs:
     - name: Perform unit tests
       run: cd build && make test
     - name: Perform code coverage
-      run: cd build && ctest -T Test -T Coverage > coverage.txt
+      run: cd build && make test && gcovr -r ../ . --xml-pretty > coverage.xml
     - name: Upload coverage result for job 1
       uses: actions/upload-artifact@v3
       with:
         name: code-coverage
         path: ./build/coverage.txt
+    - name: Code Coverage Summary Report
+      uses: irongut/CodeCoverageSummary@v1.3.0
+      with:
+        filename: coverage.xml
 
   build-ubuntu-20-04:
 
