### 2023-09-04 9e4d4c0
Adding gcovr to compile dependencies
