+++ .github/workflows/build.yml
@@ -7,7 +7,7 @@ on:
     branches: [ "master", "develop" ]
 
 jobs:
-  build-ubuntu-22.04:
+  build-ubuntu-22-04:
 
     runs-on: ubuntu-22.04
 
