+++ .github/workflows/build.yml
@@ -32,6 +32,7 @@ jobs:
       uses: irongut/CodeCoverageSummary@v1.3.0
       with:
         filename: ./build/coverage.xml
+        badge: true
 
   build-ubuntu-20-04:
 
