+++ .github/workflows/build.yml
@@ -37,6 +37,12 @@ jobs:
       with:
         filename: ./build/coverage.xml
         badge: true
+    - name: Upload coverage reports to Codecov
+      uses: codecov/codecov-action@v3
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        files: ./build/coverage.xml
+
 
   build-ubuntu-20-04:
 
