+++ .github/workflows/build.yml
@@ -26,17 +26,11 @@ jobs:
         cd build
         gcovr -r ../ . --xml-pretty > coverage.xml
         gcovr -r ../ . --html > report.html
-        tar -cvjf report.tar.bz2 report.html report.css
     - name: Upload coverage report
       uses: actions/upload-artifact@v3
       with:
         name: code-coverage-report
-        path: ./build/report.tar.bz2
-    - name: Code Coverage Summary Report
-      uses: irongut/CodeCoverageSummary@v1.3.0
-      with:
-        filename: ./build/coverage.xml
-        badge: true
+        path: ./build/report.html
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
       with:
