+++ .codecov.yml
@@ -1,12 +1,20 @@
 coverage:
   precision: 2
   round: down
-  range: "70...100"
+  range: "30...100"
 
   status:
     project: no
     patch: yes
     changes: no
+
+parsers:
+  gcov:
+    branch_detection:
+      conditional: yes
+      loop: yes
+      method: no
+      macro: no
 
 comment:
   layout: "header, diff, changes, tree"
