### 2023-09-05 1dfeb18
Adding computational times
