+++ docs/examples.rst
@@ -73,13 +73,13 @@ acceleration, we refer to the Table as seen below.
 
    * - System
      - Execution time (averaged)
-   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads)
+   * - Intel(R) Core(TM) i9-10900K (20 threads)
      - 172.58s
-   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads) + RTX 4090
+   * - Intel(R) Core(TM) i9-10900K (20 threads) + RTX 4090
      - 90.18s
-   * - Intel(R) Core(TM) i5-8400 CPU @ 2.80GHz (6 threads)
+   * - Intel(R) Core(TM) i5-8400 (6 threads)
      - 311.84s
-   * - Intel(R) Core(TM) i5-8400 CPU @ 2.80GHz (6 threads) + RTX 2070
+   * - Intel(R) Core(TM) i5-8400 (6 threads) + RTX 2070
      - 125.08s
 
 Co HCP 11-21
@@ -196,7 +196,15 @@ acceleration, we refer to the Table as seen below.
 
    * - System
      - Execution time (averaged)
-   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads)
+   * - Intel(R) Core(TM) i9-10900K (20 threads)
      - 5207.93s (1h26m47s)
-   * - Intel(R) Core(TM) i9-10900K CPU @ 3.70GHz (20 threads) + RTX 4090
+   * - Intel(R) Core(TM) i9-10900K (20 threads) + RTX 4090
      - 2368.63s (39m28s)
+   * - Intel(R) Core(TM) i5-8400 (6 threads) + RTX 2070
+     - 2986.00s (49m46s)
+   * - Intel(R) Xeon(R) Gold 6234 (16 threads) + A5000
+     - 3912.19 (65m12s)
+   * - Intel(R) Core(TM) i5-12400F (12 threads) + 1x GTX 1080 Ti
+     - 2759.49 (45m59s)
+   * - Intel(R) Core(TM) i5-12400F (12 threads) + 2x GTX 1080 Ti
+     - 2067.24s (34m27s)
