### 2023-09-06 #5710 @jgostick
Hi @ifilot. Sorry for my silence on this, I was not feeling 100% the past few weeks.  Anyway, I think we can move forward with the publication now.  @NicoRenaud was very positive, and @hsulab has checked most boxes.  There seem to be a few things to finish up first:

1. Community guidelines are missing.  These are typically added in the form of a CONTRIBUTING.md file, which lists how people can report bugs, make feature requests, and submit code.  I'm sure there is a template somewhere you can use as inspiration.
2. @hsulab did not see any speedup on the GPU and @NicoRenaud was not able to testing lacking a GPU.  You might want to clarify this aspect of the performance in the readme, but I suppose that is not critical to the overall package.
3. I think the most important thing is to look at the comments that @hsulab made about installation issues.  They had to hack around in the make file to get it working.  Is there anything you can learn from their experience to make the installation more robust?
4. Both reviewers commented on the rather large memory requirements.  Is this something that is expected, or can you fix it?
