### 2023-09-06 036f06f
Polishing CMake files
