+++ .github/workflows/build.yml
@@ -16,7 +16,7 @@ jobs:
     - name: Install dependencies
       run: sudo apt-get update && sudo apt-get install -y build-essential cmake libboost-all-dev pkg-config libeigen3-dev libtclap-dev gcovr
     - name: Configure CMake
-      run: mkdir build && cd build && cmake ../src
+      run: mkdir build && cd build && cmake ../src -DUSE_GCOV=1
     - name: Build
       run: cd build && make -j
     - name: Perform unit tests
@@ -47,7 +47,7 @@ jobs:
     - name: Install dependencies
       run: sudo apt-get update && sudo apt-get install -y build-essential cmake libboost-all-dev pkg-config libeigen3-dev libtclap-dev
     - name: Configure CMake
-      run: mkdir build && cd build && cmake ../src
+      run: mkdir build && cd build && cmake ../src -DUSE_GCOV=1
     - name: Build
       run: cd build && make -j
     - name: Perform unit tests
@@ -63,7 +63,7 @@ jobs:
     - name: Install dependencies
       run: apt-get update && apt-get install -y build-essential cmake libboost-all-dev pkg-config libeigen3-dev libtclap-dev
     - name: Configure CMake
-      run: mkdir build && cd build && cmake ../src
+      run: mkdir build && cd build && cmake ../src -DUSE_GCOV=1
     - name: Build
       run: cd build && make -j
     - name: Perform unit tests
@@ -79,7 +79,7 @@ jobs:
     - name: Install dependencies
       run: apt-get update && apt-get install -y build-essential cmake libboost-all-dev pkg-config libeigen3-dev libtclap-dev
     - name: Configure CMake
-      run: mkdir build && cd build && cmake ../src
+      run: mkdir build && cd build && cmake ../src -DUSE_GCOV=1
     - name: Build
       run: cd build && make -j
     - name: Perform unit tests
