+++ src/brambletool.cpp
@@ -69,6 +69,7 @@ int main(int argc, char* argv[]) {
         std::cout << "Documentation: https://bramble.imc-tue.nl" << std::endl;
         std::cout << "--------------------------------------------------------------" << std::endl;
         std::cout << "Compilation time: " << __DATE__ << " " << __TIME__ << std::endl;
+        std::cout << "Git Hash: " << PROGRAM_GIT_HASH << std::endl;
         #ifdef MOD_CUDA
         std::cout << "Modules: CUDA" << std::endl;
         #endif // MOD_CUDA
