+++ src/CMakeLists.txt
@@ -37,6 +37,14 @@ if(USE_GCOV)
   set(CMAKE_SHARED_LINKER_FLAGS "${CMAKE_SHARED_LINKER_FLAGS} -lgcov -profile-arcs")
 endif()
 
+# store git hash
+execute_process(
+    COMMAND git log -1 --format=%h
+    WORKING_DIRECTORY ${CMAKE_CURRENT_LIST_DIR}
+    OUTPUT_VARIABLE GIT_HASH
+    OUTPUT_STRIP_TRAILING_WHITESPACE
+)
+
 # prepare configuration file
 SET(VERSION_MAJOR "1")
 SET(VERSION_MINOR "1")
@@ -142,9 +150,3 @@ if(NOT DISABLE_TEST)
 else()
     message("[USER] Testing routine disabled")
 endif()
-
-###
-# Installing
-##
-install (TARGETS bramble DESTINATION "/opt/bramble/bin")
-install (FILES "../patterns/patterns.json" DESTINATION "/opt/bramble")
