+++ src/config.h.in
@@ -29,8 +29,10 @@
 #define VERSION_MINOR @VERSION_MINOR@
 #define VERSION_PATCH @VERSION_PATCH@
 #define VERSION "@VERSION_MAJOR@.@VERSION_MINOR@.@VERSION_PATCH@"
+#define GIT_HASH "@GIT_HASH@"
 
 static const std::string PROGRAM_VERSION(VERSION);
+static const std::string PROGRAM_GIT_HASH(GIT_HASH);
 static const unsigned int PROGRAM_VERSION_MAJOR = VERSION_MAJOR;
 static const unsigned int PROGRAM_VERSION_MINOR = VERSION_MINOR;
 static const unsigned int PROGRAM_VERSION_PATCH = VERSION_PATCH;
