### 2023-09-06 53be895
Changing CUDA compilation procedure
