### 2023-09-07 f7534ce
Adding checks for GPU
