+++ src/similarity_analysis.cpp
@@ -110,6 +110,11 @@ void SimilarityAnalysis::analyze(const std::shared_ptr<State>& _state) {
             CardManager cm;
             cuda_devices = cm.get_num_gpus();
             this->num_gpu = std::min((unsigned int)cuda_devices, this->num_gpu);
+            if(this->num_gpu > 0) {
+                if(!(cm.get_memory_device(0) >= 8589934592)) {
+                    throw std::runtime_error("Your GPU does not have at least 8Gb of memory. Cannot use this GPU.");
+                }
+            }
             std::cout << "Executing similarity analysis using " << this->num_gpu << " GPUs and " << nrthreads << " CPU threads." << std::endl;
         }
     }
