### 2023-09-07 5243a60
Fixing memory requirements
