### 2023-09-07 d8d16fb
Improving GPU usage (#11)
* Improving installation instructions

* Adding execution times

* Migrating execution times

* Adding calculation times for Co1121 example

* Add notation execution times

* Implementing multi-GPU procedure

* Adding command line arguments for NGPU

* Fixing bug in GPU number

* Changing output

* Expanding documentation

* Adding coverage report

* Adding gcovr to compile dependencies

* Specifying path

* Fixing names

* Adding badge

* Adding code coverage report

* Adding code coverage

* Changing procedure

* Modifying codecov

* Adding parsers

* Adding exclusi
[…]
