### 2023-09-13 a6f4d91
Adding Blender file and adding CONTRIBUTIONS
