### 2023-09-13 2f32710
Changing CUDA requirements (#14)
