### 2023-09-13 57c7c9a
Adding README file for scripts (#15)
* Changing CUDA requirements

* Adding README.md
