### 2023-09-13 #5710 @ifilot
I would like to thank @NicoRenaud @hsulab once more for their valuable time and effort reviewing the code and the manuscript. It is highly appreciated. The reviewer's concern are found below and we address these per point.

Review of @hsulab:
- The compilation works with eigen 3.3.4, 3.3.7, and 3.4.0 but it failed with g++ 8.2 related to std::filesystem and g++ 11.2 error: call of overloaded 'sqrt(double)' is ambiguous.
_The CUDA module requires on the availability of the ``#define EIGEN_NO_CUDA`` variable, which was only added since version 3.4.0. We have adjusted the [documentation](https://bramble.imc-tue.nl/installation.html) and mention this item._
- The reviewer needed to add `-no-pie` to `CMakeLists.txt` to compile the CUDA module
_We were unable to reproduce this error. We nevertheless adjusted the compilation instructions for CUDA and believe this implementation is more robust across different operating systems. We have tested the revised `CMakeLists.txt` file for various HPC infrastructure and workstations, all with good results._
- Also, I wonder if bramblecuda is only used for showing the GPU information and the actual computation is still performed by bramble.
_The assessment of the reviewer is correct. The only purpose of `bramblecuda` is to screen for available GPUs. We have adjusted the documentation to reflect upon this point. See [this link](https://bramble.imc-tue.nl/installation.html#cuda-support)._
- I tested the first example and did not finish the second one as it took too long. The outputs by cpu and gpu versions are shown below. I did not observe any acceleration by GPU. I wonder if my GPU version was not compiled successfully. I think the author should add some benchmarks in the documentation.
_We thank the reviewer for this valuable suggestion. We have tested `bramble` on a number of different systems and report execution times for various GPUs and CPU architectures. Tables with execution times can be found [here](https://bramble.imc-tue.nl/examples.html#execution-times) and [here](https://bramble.imc-tue.nl/examples.html#id2)._
- There is an error when reading pattern2.json.
_We thank the reviewer for bringing this error to our attention. We have corrected the file._
- Community guidelines are not provided.
_The community guidelines were only mentioned in the documentation and we forgot to also mention them in the Github repository. We thank the reviewer for bringing this to our attention. A [CONTRIBUTION.md](https://github.com/ifilot/bramble/blob/master/CONTRIBUTING.md) file has been added to the Github repository._

Review of @NicoRenaud :
- The functionality mentionned in the documentation and in the paper works well. The code seems to be rather memory demanding as the second example got killed by the OOM killer. Is it expected ?
_`bramble` requires roughly 8Gb of memory per running thread. Depending on the use case, an OOM error is possible. We have added a [page explaining the execution model](https://bramble.imc-tue.nl/execution_model.html) of `bramble` in the documentation. We also added in several places warnings mentioning the relatively large memory requirements of the program._
- I haven't been able to test the GPU performance, I have not GPU available locally.
_We have performed a - hopefully - sufficiently extensive set of tests on GPUs. We hope that the reviewer is satisfied with these results._
- The code is properly tested with GH action. It could be nice to add code coverage information in the future.
_We have added code coverage results to the Github page. See this [link](https://codecov.io/gh/ifilot/bramble)._
- In the example, it is not clear how one display the distance matrix. I guess it's with the scripts/similary_plot.py. A better documentation of this python script might help new users.'
_We thank the reviewer for this useful suggestion. We have added a reference to the script in the documentation and explain upon the working of the `similarity_plot.py` file._
- It would also be nice to add the blender script used to generate the 3d rendering of the nanocrystals.
_We have added the Blender script to the repository. Please see [this link](https://github.com/ifilot/bramble/tree/master/scripts/blender)._

Other adjustments:
- In version 1.0.0 of `bramble`, only a single GPU could be used. We have adjusted the code such that multiple GPUs can be employed using the `--ngpu` CLI. The documentation has been expanded to reflect upon this new feature.
- Previously, the user had to supply the architecture for the GPU compilation. We have now adjusted the compilation such that for a large number of common architectures the GPU code is generated. We believe this to be more user friendly.
- The review of @hsulab made us realize that the CMake route might not always be the most favorable for HPC facilities. As such, we have added a section how to compile the code using EasyBuild. See [this page](https://bramble.imc-tue.nl/installation.html#easybuild-installation).
- When a GPU does not have the required 8Gb of VRAM, the program will throw an error.

We hope that the reviewers are satisfied with these changes. We remain available for any follow-up comments or questions.
