### 2023-09-15 #5710 @hsulab
Thanks @ifilot Everything looks perfect to me.

I tested the updated bramble. The installation is smooth and the GPU acceleration is great.

The first example (bramble -s -i src/test/data/POSCAR_Rh111 -o sa_fcc111.txt) was tested on Intel(R) Xeon(R) CPU E5-2680 v4 @ 2.40GHz and Tesla P100-PCIE-16GB

```
--------------------------------------------------------------
Executing Bramble v.1.1.2
Author: Ivo Filot <i.a.w.filot@tue.nl>
Documentation: https://bramble.imc-tue.nl
--------------------------------------------------------------
Compilation time: Sep 15 2023 16:03:29
Git Hash: 57c7c9a
--------------------------------------------------------------

Loading src/test/data/POSCAR_Rh111 as POSCAR/CONTCAR file.
Geometry file contains 7 atoms.
Calculating interatomic distances...
Completed calculation of interatomic distances.
Start Similarity Analysis
Copying interatomic distances to similarity analysis object.
Constructing local density matrices... Done!
Constructing jobs and pre-allocating permutation vectors.
Building permutations: 9
Building permutations: 12
Executing similarity analysis using 28 threads.
Jobs remaining: 0
 100% [|||||||||||||||||||||||||||||||||||||||||||||||||||||]

--------------------------------------------------------------
Done in 374.825 seconds.
```

```
--------------------------------------------------------------
Executing Bramble v.1.1.2
Author: Ivo Filot <i.a.w.filot@tue.nl>
Documentation: https://bramble.imc-tue.nl
--------------------------------------------------------------
Compilation time: Sep 15 2023 16:06:39
Git Hash: 57c7c9a
Modules: CUDA
--------------------------------------------------------------

Loading src/test/data/POSCAR_Rh111 as POSCAR/CONTCAR file.
Geometry file contains 7 atoms.
Calculating interatomic distances...
Completed calculation of interatomic distances.
Start Similarity Analysis
Copying interatomic distances to similarity analysis object.
Constructing local density matrices... Done!
Constructing jobs and pre-allocating permutation vectors.
Building permutations: 9
Building permutations: 12
Executing similarity analysis using 1 GPUs and 28 CPU threads.
Jobs remaining: 0
 100% [|||||||||||||||||||||||||||||||||||||||||||||||||||||]

--------------------------------------------------------------
Done in 223.139 seconds.
```

All the best,
Jiayan
