### 2023-09-16 #5710 @jgostick
@hsulab and @NicoRenaud, I have taken the liberty of updating your checklists based on the above comments about how you're both satisfied.  (The checklists are the easiest way for me to keep track of things!).  I wish there was a way I could thank you for your efforts here other than just saying "thank you"!  You've both done an outstanding job, and it is essential for this process, so "thank you".

@ifilot, congratulations, your package has passed the review!  There are still a number of steps we need to take care of before we can move to 'accept'.  I will create another "post review" checklist which has all the task for you and I to do.
