### 2023-09-16 #5710 @jgostick
## References
- [ ] Is it normal to include the company (NVIDIA) as an author like this "NVIDIA, Vingelmann, P., & Fitzek, F. H. P. (2020)"?  Also, the link is to a rather generic website.  Is there a page or publication about this package?
- [ ] In the text there is inconsistent spacing between the name and the citation, like Jones et[cite] and Smith et al [cite].  Also, sometimes the [cite] comes after the period but not others.  I have scanned the JOSS docs and cannot find an official style guide so I'll just ask that the authors pick their preferred style and be consistent with it.
