### 2023-09-16 #5710 @ifilot
I checked the ORCIDs and these are in order.
I made a new release. The version to be used is: **v1.1.2** [link](https://github.com/ifilot/bramble/releases/tag/v1.1.2)
The software has been archied and the Zenodo DOI is: **10.5281/zenodo.8341189** [link](https://zenodo.org/record/8341189)
The title and author list matches those in the JOSS paper.
The license in the Zenodo archive corresponds to the one mentioned on Github. (GPLv3)
