### 2023-09-16 #5710 @jgostick
Hi @ifilot can you address [my two comments](https://github.com/openjournals/joss-reviews/issues/5710#issuecomment-1722091173) about the reference/citation format in the paper?
