### 2023-09-16 #5710 @ifilot
@jgostick Most certainly! (I was already working on it)

No problem addressing the second point, though I have a question about the first point. I was looking into the best way to cite the CUDA toolkit and according to [this source](http://citebay.com/how-to-cite/cuda/), it is recommended to cite the website. That same source also offers to cite the textbook (see below). I have no strong preference myself; would you prefer the textbook?

```
@book{10.5555/2430671,
author = {Cook, Shane},
title = {CUDA Programming: A Developer’s Guide to Parallel Computing with GPUs},
year = {2012},
isbn = {9780124159334},
publisher = {Morgan Kaufmann Publishers Inc.},
address = {San Francisco, CA, USA},
edition = {1st}
}
```
