### 2023-09-16 #5710 @jgostick
I also have no preference, but thought that including NVIDIA the company as an author seemed odd. Maybe let's go with the textbook since it'll be easier to find in 5 years than a website that will almost certainly die of link rot eventually.
