+++ paper/paper.bib
@@ -166,9 +166,12 @@
   year      = {2007}
 }
 
-@misc{cuda,
-  author={NVIDIA and Vingelmann, Péter and Fitzek, Frank H.P.},
-  title={CUDA, release: 10.2.89},
-  year={2020},
-  url={https://developer.nvidia.com/cuda-toolkit},
+@book{cuda,
+    author = {Cook, Shane},
+    title = {CUDA Programming: A Developer’s Guide to Parallel Computing with GPUs},
+    year = {2012},
+    isbn = {9780124159334},
+    publisher = {Morgan Kaufmann Publishers Inc.},
+    address = {San Francisco, CA, USA},
+    edition = {1st}
 }
