### 2023-09-16 534b39e
Paper (#16)
* Fixing reference placement and adding reference

* Changing CUDA reference
