+++ paper/paper.bib
@@ -55,6 +55,17 @@
     doi = {10.1021/jacsau.2c00430}
 }
 
+@article{vanetten:2023,
+    author = {van Etten, Michel P. C. and De Laat, M.E. and Hensen, Emiel J. M. and Filot, Ivo A. W.},
+    title = {Unraveling the Role of Metal–Support Interactions on the Structure Sensitivity of Fischer–Tropsch Synthesis},
+    journal = {J. Phys. Chem. C},
+    volume = {127},
+    number = {31},
+    pages = {15148-15156},
+    year = {2023},
+    doi = {10.1021/acs.jpcc.3c02240},
+}
+
 @article{reinhart:2017,
     author = {Reinhart, Wesley F. and Long, Andrew W. and Howard, Michael P. and Ferguson, Andrew L. and Panagiotopoulos, Athanassios Z.},
     title  = {Machine learning for autonomous crystal structure identification},
@@ -155,9 +166,12 @@
   year      = {2007}
 }
 
-@misc{cuda,
-  author={NVIDIA and Vingelmann, Péter and Fitzek, Frank H.P.},
-  title={CUDA, release: 10.2.89},
-  year={2020},
-  url={https://developer.nvidia.com/cuda-toolkit},
+@book{cuda,
+    author = {Cook, Shane},
+    title = {CUDA Programming: A Developer’s Guide to Parallel Computing with GPUs},
+    year = {2012},
+    isbn = {9780124159334},
+    publisher = {Morgan Kaufmann Publishers Inc.},
+    address = {San Francisco, CA, USA},
+    edition = {1st}
 }
