### 2023-09-16 #5710 @ifilot
@jgostick The paper has been revised, please find below a list of changes:

- Citations have been made more consistently. (1) All citations follow after a space. (2) Citations that would be at the end of a sentence are placed after the period (".").
- The citation to CUDA has been changed from the website to the textbook.
- One additional citation has been added corresponding to another work using the Bramble tool that got recently accepted: [10.1021/acs.jpcc.3c02240](https://doi.org/10.1021/acs.jpcc.3c02240)
