### 2023-09-26 #5710 @Kevin-Mattheus-Moerman
@jgostick thanks for editing, I've started to process this now. Note for the future, we tend to make the version tag for JOSS match the string listed for the tagged release on GitHub (https://github.com/ifilot/bramble/releases/tag/v1.1.2) which here includes the `v`. That then should also match the ZENODO version tag (it does so already so that is good).
