### 2023-09-26 #5710 @ifilot
@Kevin-Mattheus-Moerman Thank you for spotting this missing item! I have updated the source files in the repository and the city and country have been added to the affiliation.
