### 2023-09-26 #5710 @Kevin-Mattheus-Moerman
@ifilot congratulations on this publication in JOSS

@jgostick thanks for editing!

And, a special thanks to the reviewers: @hsulab, @NicoRenaud !!
